move logic from action plugin to command module
